Always quote ANONYMIZE_REMOTE_ADDR
The environment var contains a truthiness value, so the YAML/JSON parser will detect it as boolean and will not emit quotes around the value. Newer K8s version require env vars to be strings (or rather: they check to see if they actually are of type string), so when the template renders with true or false it will complain about that not being a string.